LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/static-router-request-method.https-expected.txt


FAIL Subresource load matched with the requestMethod GET condition assert_equals: expected "Network with GET request" but got "nlgjk"
FAIL Subresource load matched with the requestMethod POST condition assert_equals: expected "Network with POST request" but got "mdfhk"
FAIL Subresource load matched with the requestMethod PUT condition assert_equals: expected "Network with PUT request" but got "vmhwc"
FAIL Subresource load matched with the requestMethod DELETE condition assert_equals: expected "Network with DELETE request" but got "mtvce"
